### 2023-10-30 #136 @whedon
[PRE REVIEW]: Pigeons.jl: Distributed sampling from intractable distributions
**Submitting author:** @nikola-sur (<a href="http://orcid.org/0000-0002-5629-7910">Nikola Surjanovic</a>)
**Repository:** <a href="https://github.com/Julia-Tempering/Pigeons-Paper" target ="_blank">https://github.com/Julia-Tempering/Pigeons-Paper</a>
**Version:**
**Managing EiC:** Mathieu Besançon
