### 2023-10-30 5fa4b44
Remove inputs and place in main tex
