+++ src/PT.tex
@@ -1,306 +1,0 @@
-\section{Parallel tempering}
-\label{sec:PT}
-Pigeons.jl provides an implementation of distributed parallel tempering (PT) described 
-in \cite{syed2021nrpt}, which we outline in Algorithm~\ref{alg:distributed_PT}.
-This section gives both a brief overview of PT and some details of 
-its distributed implementation.
-
-\medskip 
-In this section we assume a basic understanding of 
-Markov chain Monte Carlo (MCMC) methods.
-For readers unfamiliar with MCMC, it is important to know that 
-it is a method to obtain approximate samples from a distribution. 
-More specifically, it is a class of algorithms for simulating a Markov chain of 
-states that look like draws from the target distribution $\pi$.
-However, with traditional 
-MCMC methods, the samples may be heavily correlated instead of independent 
-and may fail to sufficiently explore the 
-full space of the distribution (see \cref{fig:bimodal}); 
-PT is a method that aims to address these issues.
-For a more in-depth review of PT, we refer readers to \cite{surjanovic2022vpt}.
-
-
-\subsection{Overview of PT}
-Suppose that we would like to estimate integrals involving $\pi$, such as 
-$\int f(x) \pi(x) \, \dee x$. 
-These integrals may be multivariate and even include combinations of 
-continuous and discrete variables (where sums replace integrals in the discrete 
-case).
-One method is to obtain samples from $\pi$ to approximate such integrals.
-Often, the distribution 
-$\pi$ can be challenging for traditional MCMC methods---such 
-as Metropolis-Hastings, slice sampling, and Hamiltonian Monte Carlo---
-because of its structure.
-For example, in a bimodal example such as the one illustrated in 
-Figure~\ref{fig:bimodal},
-traditional methods might remain in one of the two modes for an extremely 
-long period of time.
-
-\begin{figure}[t]
-    \centering
-    \includegraphics[width=0.3\textwidth]{img/bimodal.pdf}
-    \caption{A simple bimodal distribution from which traditional 
-    MCMC methods may struggle to obtain samples. Blue lines display output 
-    from 1,000 iterations of a Metropolis-Hastings random walk MCMC algorithm.
-    The sampler in this figure is visibly trapped in one of the two modes.}
-    \label{fig:bimodal}
-\end{figure}
-
-To resolve this issue, PT constructs a sequence of $N$ distributions,  
-$\pi_1, \pi_2, \ldots, \pi_N$, where $\pi_N$ is usually equal to $\pi$.
-The distributions are chosen so that it is easy to obtain samples from $\pi_1$
-with the sampling difficulty increasing as one approches $\pi_N$. An example 
-of such a sequence of distributions, referred to as an \textit{annealing path},
-is shown in Figure~\ref{fig:path}.
-
-\begin{figure}[t]
-    \centering
-    \begin{subfigure}{0.15\textwidth}
-      \centering
-      \includegraphics[width=\textwidth]{img/heatmap_path_1.pdf}
-      \caption*{$\pi_1$}
-    \end{subfigure}
-    \begin{subfigure}{0.15\textwidth}
-      \centering
-      \includegraphics[width=\textwidth]{img/heatmap_path_2.pdf}
-      \caption*{$\pi_2$}
-    \end{subfigure}
-    \begin{subfigure}{0.15\textwidth}
-      \centering
-      \includegraphics[width=\textwidth]{img/heatmap_path_3.pdf}
-      \caption*{$\pi_3$}
-    \end{subfigure}
-    \begin{subfigure}{0.15\textwidth}
-      \centering
-      \includegraphics[width=\textwidth]{img/heatmap_path_4.pdf}
-      \caption*{$\pi_4$}
-    \end{subfigure}
-    \begin{subfigure}{0.15\textwidth}
-      \centering
[…]

[…]
