### 2023-10-30 7e7e0b7
Revert "Remove inputs and place in main tex"
This reverts commit 5fa4b448d674f0740c463b4df871991756a7c766.
