+++ src/examples.tex
@@ -1,0 +1,438 @@
+\section{Examples}
+In this section we present a set of minimal examples 
+that demonstrate how to use Pigeons.jl for sampling. 
+For further reading, we also direct readers to our growing list of examples, which can be found at 
+\url{https://github.com/Julia-Tempering/Pigeons.jl/tree/main/examples}.
+We begin by installing the latest official release of Pigeons.jl: 
+\begin{lstlisting}[language = Julia]
+using Pkg; Pkg.add("Pigeons")
+\end{lstlisting}
+
+
+\subsection{Targets}
+To use Pigeons.jl, we must specify a target distribution, given by $\gamma$ 
+in \cref{eq:normalizing_constant}.
+Numerous possible types of target distributions are supported, including 
+custom probability densities (specified up to a normalizing constant) written in Julia.
+We also allow to interface with models written in common probabilistic programming 
+languages, including:
+\begin{itemize}
+    \item Turing.jl \cite{ge2018turing} models (\texttt{TuringLogPotential})
+    \item Stan \cite{carpenter2017stan} models (\texttt{StanLogPotential})
+    \item Comrade.jl\footnote{\url{https://github.com/ptiede/Comrade.jl}} 
+      models for black hole imaging (\texttt{ComradeLogPotential})
+    \item Non-Julian models with foreign-language Markov chain Monte Carlo (MCMC) code 
+    (e.g. Blang \cite{bouchard2022blang} code for phylogenetic inference over combinatorial spaces) 
+\end{itemize}
+Additional targets are currently being accommodated and will be 
+introduced to Pigeons.jl in the near future.
+
+\medskip 
+In what follows, we demonstrate how to use Pigeons with a Julia Turing model 
+applied to a non-identifiable ``coinflip'' data set, modified from the example
+at \url{https://turing.ml/v0.22/docs/using-turing/quick-start}. 
+The Bayesian model can be formulated as 
+\[
+  \label{eq:coinflip}
+  p_1, p_2 &\stackrel{iid}{\sim} U(0, 1) \\    
+  Y \mid p_1, p_2 &\sim \text{Binomial}(n, p_1 p_2).
+\]
+The random variable $Y$ is the number of heads observed on $n$ coin flips
+where the probability of heads is $p_1 p_2$.  
+This model is non-identifiable, meaning that it is not possible to distinguish 
+the effects of the two different parameters $p_1$ and $p_2$. As a consequence, 
+the target distribution exhibits a complicated structure, as displayed in 
+\cref{fig:coinflip_posterior}.
+The density of interest corresponding to this model is 
+\[
+  \pi(p_1, p_2) &= \gamma(p_1, p_2)/Z,
+\] 
+where 
+\[ 
+  \gamma(p_1, p_2) &= 
+    {n \choose y} (p_1 p_2)^y (1-p_1 p_2)^{n-y} I[p_1, p_2 \in [0,1]] 
+    \label{eq:coinflip_density} \\
+  Z &= \int_0^1 \int_0^1 \gamma(p_1, p_2) \, \dee p_1 \, \dee p_2. \label{eq:coinflip_normalization}
+\]
+The distribution $\pi$ is also known as the \emph{posterior distribution} in 
+Bayesian statistics.
+
+\begin{figure}[t]
+    \centering 
+    \includegraphics[width=0.5\textwidth]{img/coinflip_posterior.pdf}
+    \caption{Posterior distribution for the model given by \cref{eq:coinflip} 
+    with $n=100,000$ coin flips and $y=50,000$ observed heads, 
+    estimated using $2^{17}$ samples from Pigeons.jl. 
+    We present the pairwise plot for $p_1$ and $p_2$, as well as the estimated 
[…]

[…]
