+++ paper.tex
@@ -47,7 +47,6 @@ a distributed and randomized algorithm that satisfies strong parallelism invaria
 \input{src/PT.tex}
 \input{src/SPI.tex}
 \input{src/related.tex}
-% \input{src/benchmarks.tex}
 \input{src/conclusion.tex}
 
 \section{Acknowledgements}
