+++ src/benchmarks.tex
@@ -1,2 +1,0 @@
-\section{Benchmarks}
-We perform various simple experiments to benchmark ...
