+++ paper.tex
@@ -42,12 +42,12 @@ a distributed and randomized algorithm that satisfies strong parallelism invaria
 
 \end{abstract}
 
-\input{src/introduction.tex}
-\input{src/examples.tex}
-\input{src/PT.tex}
-\input{src/SPI.tex}
-\input{src/related.tex}
-\input{src/conclusion.tex}
+% \input{src/introduction.tex}
+% \input{src/examples.tex}
+% \input{src/PT.tex}
+% \input{src/SPI.tex}
+% \input{src/related.tex}
+% \input{src/conclusion.tex}
 
 \section{Acknowledgements}
 NS acknowledges the support of a Vanier Canada Graduate Scholarship. 
