+++ ref.bib
@@ -60,7 +60,8 @@
   volume={23},
   pages={535--549},
   year={2013},
-  publisher={Springer}
+  publisher={Springer}, 
+  doi = {10.1007/s11222-012-9328-6}
 }
 
 @article{gilks1994language,
@@ -71,7 +72,8 @@
   number={1},
   pages={169--177},
   year={1994},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library}, 
+  doi = {10.2307/2348941}
 }
 
 @article{goudie2020multibugs,
@@ -105,7 +107,8 @@
   number={4},
   pages={611--615},
   year={2013},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press}, 
+  doi = {10.1093/sysbio/syt022}
 }
 
 @article{martin2022korali,
@@ -116,7 +119,8 @@ pages = {114264},
 year = {2022},
 issn = {0045-7825},
 url = {https://www.sciencedirect.com/science/article/pii/S0045782521005752},
-author = {Sergio M. Martin and Daniel Wälchli and Georgios Arampatzis and Athena E. Economides and Petr Karnakov and Petros Koumoutsakos}
+author = {Sergio M. Martin and Daniel Wälchli and Georgios Arampatzis and Athena E. Economides and Petr Karnakov and Petros Koumoutsakos}, 
+doi = {10.1016/j.cma.2021.114264}
 }
 
 @article{langmore2023hamiltonian,
@@ -127,7 +131,8 @@ author = {Sergio M. Martin and Daniel Wälchli and Georgios Arampatzis and Athen
 	year    = {2023},
 	volume  = {13},
 	number  = {1},
-	pages   = {69--93}
+	pages   = {69--93}, 
+  doi = {10.1615/int.j.uncertaintyquantification.2022038478}
 }
 
 @article{Eirew2022,
@@ -139,7 +144,8 @@ author = {Sergio M. Martin and Daniel Wälchli and Georgios Arampatzis and Athen
   number = {1},
   author = {Peter Eirew and Ciara O'Flanagan and Jerome Ting and Sohrab Salehi and Jazmine Brimhall and Beixi Wang and Justina Biele and Teresa Algara and So Ra Lee and Corey Hoang and Damian Yap and Steven McKinney and Cherie Bates and Esther Kong and Daniel Lai and Sean Beatty and Mirela Andronescu and Elena Zaikova and Tyler Funnell and Nicholas Ceglia and Stephen Chia and Karen Gelmon and Colin Mar and Sohrab Shah and Andrew Roth and Alexandre Bouchard-C{\^{o}}t{\'{e}} and Samuel Aparicio},
   title = {Accurate determination of {CRISPR}-mediated gene fitness in transplantable tumours},
-  journal = {Nature Communications}
+  journal = {Nature Communications}, 
+  doi = {10.1038/s41467-022-31830-2}
 }
 
 
@@ -184,7 +190,8 @@ author = {Sergio M. Martin and Daniel Wälchli and Georgios Arampatzis and Athen
   author={Cusumano-Towner, Marco F and Saad, Feras A and Lew, Alexander K and Mansinghka, Vikash K},
   booktitle={Proceedings of the 40th ACM SIGPLAN Conference on Programming Language Design and Implementation},
   pages={221--236},
-  year={2019}
+  year={2019}, 
+  doi = {10.1145/3314221.3314642}
 }
 
 @article{hoffman2014no,
@@ -216,7 +223,8 @@ author = {Sergio M. Martin and Daniel Wälchli and Georgios Arampatzis and Athen
   number={4},
   pages={327--335},
   year={1995},
-  publisher={Taylor \& Francis}
[…]
