+++ paper.tex
@@ -4,6 +4,7 @@
 
 \usepackage[autonum, blackhypersetup]{shortex}
 \usepackage{xcolor}
+\let\subcaption\undefined % overwrite 'subcaption' from the style file
 \usepackage[font=scriptsize]{subcaption}
 \usepackage{forest}
 
