+++ header.tex
@@ -1,23 +1,18 @@
 % **************GENERATED FILE, DO NOT EDIT**************
 
-\title{Pigeons.jl: Distributed sampling from intractable distributions}
+\title{My JuliaCon proceeding}
 
-\author[1]{Nikola Surjanovic}
-\author[1]{Miguel Biron-Lattes}
-\author[2]{Paul Tiede}
-\author[3]{Saifuddin Syed}
-\author[1]{Trevor Campbell}
-\author[1]{Alexandre Bouchard-Côté}
-\affil[1]{University of British Columbia}
-\affil[2]{Harvard University}
-\affil[3]{University of Oxford}
+\author[1]{1st author}
+\author[1, 2]{2nd author}
+\author[2]{3rd author}
+\affil[1]{University}
+\affil[2]{National Lab}
 
-\keywords{Distributed computation, Bayesian inference, parallelism invariance, 
-multi-threading, Message Passing Interface, Markov chain Monte Carlo} 
+\keywords{Julia, Optimization, Game theory, Compiler}
 
 \hypersetup{
 pdftitle = {My JuliaCon proceeding},
-pdfsubject = {JuliaCon 2019 Proceedings},
+pdfsubject = {JuliaCon 2022 Proceedings},
 pdfauthor = {1st author, 2nd author, 3rd author},
 pdfkeywords = {Julia, Optimization, Game theory, Compiler},
 }
