### 2023-10-30 #136 @nikola-sur
@miguelbiron Any insight into how to view the log?
