+++ paper.tex
@@ -4,7 +4,7 @@
 
 \usepackage[autonum, blackhypersetup]{shortex}
 \usepackage{xcolor}
-\let\subcaption\undefined % overwrite 'subcaption' from the style file
+%\let\subcaption\undefined % overwrite 'subcaption' from the style file
 \usepackage[font=scriptsize]{subcaption}
 \usepackage{forest}
 
@@ -20,6 +20,8 @@
 \definecolor{julia6}{HTML}{00aaae}
 \definecolor{julia7}{HTML}{ed5e93}
 \definecolor{julia8}{HTML}{c68225}
+
+\pdfsuppresswarningpagegroup=1
 
 \begin{document}
 
