+++ src/examples.tex
@@ -50,7 +50,7 @@ The density of interest corresponding to this model is
 where 
 \[ 
   \gamma(p_1, p_2) &= 
-    {n \choose y} (p_1 p_2)^y (1-p_1 p_2)^{n-y} I[p_1, p_2 \in [0,1]] 
+    \binom{n}{y} (p_1 p_2)^y (1-p_1 p_2)^{n-y} I[p_1, p_2 \in [0,1]] 
     \label{eq:coinflip_density} \\
   Z &= \int_0^1 \int_0^1 \gamma(p_1, p_2) \, \dee p_1 \, \dee p_2. \label{eq:coinflip_normalization}
 \]
