+++ paper/PT.tex
@@ -38,7 +38,7 @@ long period of time.
 
 \begin{figure}[t]
     \centering
-    \includegraphics[width=0.3\textwidth]{img/bimodal.pdf}
+    \includegraphics[width=0.3\textwidth]{../img/bimodal.pdf}
     \caption{A simple bimodal distribution from which traditional 
     MCMC methods may struggle to obtain samples. Blue lines display output 
     from 1,000 iterations of a Metropolis-Hastings random walk MCMC algorithm.
@@ -55,36 +55,36 @@ is shown in Figure~\ref{fig:path}.
 
 \begin{figure}[t]
     \centering
-    \begin{subfigure}{0.15\textwidth}
+    \begin{minipage}{0.15\textwidth}
       \centering
-      \includegraphics[width=\textwidth]{img/heatmap_path_1.pdf}
+      \includegraphics[width=\textwidth]{../img/heatmap_path_1.pdf}
       \caption*{$\pi_1$}
-    \end{subfigure}
-    \begin{subfigure}{0.15\textwidth}
+    \end{minipage}
+    \begin{minipage}{0.15\textwidth}
       \centering
-      \includegraphics[width=\textwidth]{img/heatmap_path_2.pdf}
+      \includegraphics[width=\textwidth]{../img/heatmap_path_2.pdf}
       \caption*{$\pi_2$}
-    \end{subfigure}
-    \begin{subfigure}{0.15\textwidth}
+    \end{minipage}
+    \begin{minipage}{0.15\textwidth}
       \centering
-      \includegraphics[width=\textwidth]{img/heatmap_path_3.pdf}
+      \includegraphics[width=\textwidth]{../img/heatmap_path_3.pdf}
       \caption*{$\pi_3$}
-    \end{subfigure}
-    \begin{subfigure}{0.15\textwidth}
+    \end{minipage}
+    \begin{minipage}{0.15\textwidth}
       \centering
-      \includegraphics[width=\textwidth]{img/heatmap_path_4.pdf}
+      \includegraphics[width=\textwidth]{../img/heatmap_path_4.pdf}
       \caption*{$\pi_4$}
-    \end{subfigure}
-    \begin{subfigure}{0.15\textwidth}
+    \end{minipage}
+    \begin{minipage}{0.15\textwidth}
       \centering
-      \includegraphics[width=\textwidth]{img/heatmap_path_5.pdf}
+      \includegraphics[width=\textwidth]{../img/heatmap_path_5.pdf}
       \caption*{$\pi_5$}
-    \end{subfigure}
-    \begin{subfigure}{0.15\textwidth}
+    \end{minipage}
+    \begin{minipage}{0.15\textwidth}
       \centering
-      \includegraphics[width=\textwidth]{img/heatmap_path_6.pdf}
+      \includegraphics[width=\textwidth]{../img/heatmap_path_6.pdf}
       \caption*{$\pi_6$}
-    \end{subfigure}
+    \end{minipage}
     \caption{Heatmaps of six distributions lying on an annealing path 
     from a unimodal reference distribution  
     $\pi_1$, from which it is straightforward to obtain samples, and ending at 
@@ -191,26 +191,26 @@ step an index of the array is manipulated by exactly one machine.
 
 \begin{figure*}[t] 
   \centering 
-  \begin{subfigure}{0.48\textwidth}
+  \begin{minipage}{0.48\textwidth}
     \centering
-    \includegraphics[width=0.39\textwidth]{img/Communication_1_v3.pdf}
-    \includegraphics[width=0.59\textwidth]{img/index_process_1.pdf}
-  \end{subfigure}
-  \begin{subfigure}{0.48\textwidth}
[…]
