+++ Project.toml
@@ -1,12 +1,0 @@
-[deps]
-CairoMakie = "13f3f980-e62b-5c42-98c6-ff1f3baf88f0"
-DynamicPPL = "366bfd00-2699-11ea-058f-f148b4cae6d8"
-LogExpFunctions = "2ab3a3ac-af41-5b50-aa03-7779005ae688"
-MCMCChains = "c7f686f2-ff18-58e9-bc7b-31028e88f75d"
-PairPlots = "43a3c2be-4208-490b-832a-a21dcd55d7da"
-Pigeons = "0eb8d820-af6a-4919-95ae-11206f830c31"
-PlotlyJS = "f0f68f2c-4968-5e81-91da-67840de0976a"
-Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
-SplittableRandoms = "8efc31e9-3fb0-4277-b18c-5a3d5d07abad"
-StatsPlots = "f3b207a7-027a-5e70-b257-86293d7955fd"
-Turing = "fce5fe82-541a-59a6-adf8-730c64b5f9a0"
