### 2023-10-31 #136 @miguelbiron
Hi @pitsianis -- I'm running out of ideas here. Do you know if there's a way to print the log to debug the error?
