### 2023-10-31 #136 @pitsianis
I had requested the same thing https://github.com/JuliaCon/proceedings-review/issues/126#issuecomment-1784690548

@arfon and/or @xuanxu, please also check the logs on this submission and enlighten us. Perhaps if we had the compilation script, we would not need to bother you this often 😄
