+++ paper/header.tex
@@ -1,19 +1,23 @@
 % **************GENERATED FILE, DO NOT EDIT**************
 
-\title{My JuliaCon proceeding}
+\title{Pigeons.jl: Distributed sampling from intractable distributions}
 
-\author[1]{1st author}
-\author[1, 2]{2nd author}
-\author[2]{3rd author}
-\affil[1]{University}
-\affil[2]{National Lab}
+\author[1]{Nikola Surjanovic}
+\author[1]{Miguel Biron-Lattes}
+\author[2]{Paul Tiede}
+\author[3]{Saifuddin Syed}
+\author[1]{Trevor Campbell}
+\author[1]{Alexandre Bouchard-Côté}
+\affil[1]{University of British Columbia}
+\affil[2]{Harvard University}
+\affil[3]{University of Oxford}
 
-\keywords{Julia, Optimization, Game theory, Compiler}
+\keywords{distributed computation, Bayesian inference, parallelism invariance, multi-threading, Message Passing Interface, Markov chain Monte Carlo}
 
 \hypersetup{
-pdftitle = {My JuliaCon proceeding},
+pdftitle = {Pigeons.jl: Distributed sampling from intractable distributions},
 pdfsubject = {JuliaCon 2022 Proceedings},
-pdfauthor = {1st author, 2nd author, 3rd author},
-pdfkeywords = {Julia, Optimization, Game theory, Compiler},
+pdfauthor = {Nikola Surjanovic, Miguel Biron-Lattes, Paul Tiede, Saifuddin Syed, Trevor Campbell, Alexandre Bouchard-Côté},
+pdfkeywords = {distributed computation, Bayesian inference, parallelism invariance, multi-threading, Message Passing Interface, Markov chain Monte Carlo},
 }
 
