+++ .github/workflows/draft-pdf.yml
@@ -12,7 +12,7 @@ jobs:
         with:
           journal: joss
           # This should be the path to the paper within your repo.
-          paper-path: paper/paper.tex
+          paper-path: paper.tex
       - name: Upload
         uses: actions/upload-artifact@v1
         with:
@@ -20,5 +20,5 @@ jobs:
           # This is the output path where Pandoc will write the compiled
           # PDF. Note, this should be the same directory as the input
           # paper.md
-          path: paper/paper.pdf
+          path: paper.pdf
 
