+++ paper/paper.yml
@@ -7,7 +7,6 @@ keywords:
   - Message Passing Interface 
   - Markov chain Monte Carlo
 authors:
-
   - name: Nikola Surjanovic 
     affiliation: 1
   - name: Miguel Biron-Lattes 
