### 2023-10-31 #136 @miguelbiron
Ok so i setup the GH action and at least I can [produce an error](https://github.com/Julia-Tempering/Pigeons-Paper/actions/runs/6712653886/job/18242665302#step:4:15)....
