### 2023-11-01 #136 @arfon
As I mentioned over in https://github.com/JuliaCon/proceedings-review/issues/126#issuecomment-1784690548, we're in the process of moving to a new bot for this stuff. I'm afraid the current functionality of Whedon is essentially frozen in time right now (no new features).

Did you verify that you can run the build script yourself as documented here?
https://github.com/JuliaCon/JuliaConSubmission.jl?tab=readme-ov-file#build-process
