### 2023-11-01 #136 @pitsianis
@miguelbiron @nikola-sur replace the `paper/` directory content with the corresponding one from [here](https://github.com/JuliaCon/JuliaConSubmission.jl), it should work. Then substitute back, file by file, your sources. If the error is in `paper.tex`, do a "binary search" by commenting out parts of your text. This is the old-fashioned, binary debugging until we move to the better editorial bot.
