### 2023-11-01 #136 @miguelbiron
@arfon @pitsianis That sounds like what we have basically done so far with @nikola-sur . We could try again I guess.

I actually was able to install [whedon cmd line utils](https://github.com/openjournals/whedon) on my computer and could successfully compile our paper: [10.21105.jcon.00136.pdf](https://github.com/JuliaCon/proceedings-review/files/13230266/10.21105.jcon.00136.pdf). So all the evidence points to the current deployment of the @ whedon bot having a corrupt texlive installation.
