+++ paper/paper.yml
@@ -17,7 +17,7 @@ authors:
     affiliation: 3
   - name: Trevor Campbell
     affiliation: 1
-  - name: Alexandre Bouchard-Côté
+  - name: Alexandre Bouchard-C\^{o}t\'{e}
     affiliation: 1
 affiliations:
   - name: University of British Columbia
