+++ paper/header.tex
@@ -7,7 +7,7 @@
 \author[2]{Paul Tiede}
 \author[3]{Saifuddin Syed}
 \author[1]{Trevor Campbell}
-\author[1]{Alexandre Bouchard-Côté}
+\author[1]{Alexandre Bouchard-C\^{o}t\'{e}}
 \affil[1]{University of British Columbia}
 \affil[2]{Harvard University}
 \affil[3]{University of Oxford}
@@ -17,7 +17,7 @@
 \hypersetup{
 pdftitle = {Pigeons.jl: Distributed sampling from intractable distributions},
 pdfsubject = {JuliaCon 2022 Proceedings},
-pdfauthor = {Nikola Surjanovic, Miguel Biron-Lattes, Paul Tiede, Saifuddin Syed, Trevor Campbell, Alexandre Bouchard-Côté},
+pdfauthor = {Nikola Surjanovic, Miguel Biron-Lattes, Paul Tiede, Saifuddin Syed, Trevor Campbell, Alexandre Bouchard-C\^{o}t\'{e}},
 pdfkeywords = {distributed computation, Bayesian inference, parallelism invariance, multi-threading, Message Passing Interface, Markov chain Monte Carlo},
 }
 
