+++ paper/paper.tex
@@ -2,6 +2,7 @@
 \documentclass{juliacon}
 \setcounter{page}{1}
 
+\usepackage{expl3} % add missing functionality in the old texlive version running on the server 
 \usepackage[autonum, blackhypersetup]{shortex}
 \usepackage{xcolor}
 %\let\subcaption\undefined % overwrite 'subcaption' from the style file
