### 2023-11-18 #139 @whedon
[REVIEW]: Pigeons.jl: Distributed sampling from intractable distributions
**Submitting author:** @nikola-sur (<a href="http://orcid.org/0000-0002-5629-7910">Nikola Surjanovic</a>)
**Repository:** https://github.com/Julia-Tempering/Pigeons-Paper
**Branch with paper.md** (empty if default branch):
**Version:** Pigeons v0.4.8
