### 2023-12-10 #139 @nsailor
This work is an implementation of the Parallel Tempering method in Julia, an improvement over MCMC for obtaining samples form otherwise difficult to sample probability distributions.
The package supports parallelization using both Julia's built-in multi-threading features, as well as MPI.jl, a Julia wrapper for MPI.

I greatly appreciate the attention given to what the authors term "strong parallelism invariance" (SPI), roughly meaning that the result should not depend on the degree of parallelism (especially considering the complexities of finite-precision floating-point arithmetic). This package also builds upon a Julia translation of a Java library called Splittable Random previously published by the authors, allowing deterministic random number generation across multiple threads.

Overall, I think this work is a great contribution to the Julia ecosystem, notwithstanding the following minor points:

 * Although definitely not a requirement, I think it would be very interesting to see the scaling characteristics of this implementation, especially with respect to the other algorithm parameters, for instance, the number of chains.
Additionally, the paper mentions Pigeons being able to run on "thousands of MPI-communicating machines". It may be worth clarifying if this is something that has been tested or a possibility given the package's design. Having said that, I was glad to find [this issue](https://github.com/Julia-Tempering/Pigeons.jl/issues/179) so it seems this is something on the authors' radar.

 * Unfortuantely, I was not able to get a speedup with the toy MVN example using multiple threads (see [this issue](https://github.com/Julia-Tempering/Pigeons.jl/issues/191) in the project repository). In general, it would be very helpful to have some hints in the package's documentation for tuning the degree of parallelism for a given problem, especially if the speedup from parallelism is not linear.

 * It would be great to have additional examples, but this is also part of [this issue](https://github.com/Julia-Tempering/Pigeons.jl/issues/179)

In writing the above, it is possible that I have misunderstood some part of this work, in which case, please feel free to correct me.
