### 2024-02-27 #139 @georgebisbas
First, I would like to thank the authors for their patience over the last few months.

Pigeons.jl offers a high-level API to leverage shared- and distributed-memory parallelism via Julia's built-in multi-threading features and MPI.jl, a Julia wrapper for MPI.
Overall, this is a great work suitable for JuliaCon proceedings.
My review below lists a few weaknesses/questions that could help clarify my understanding of the work in a few places or act as constructive feedback to improve this work.

Strengths :
- Automated deployment to MPI DMP without necessary user experience
- Strong parallelism invariance is guaranteed.
- Applications are discussed in the Introduction.
- Smooth transition for the reader, problem formulation, and running examples help in that.
- Compatibility with other targets is a plus
- Section 2.2.1 helps to describe Figure 2
- I very much like how code is accompanied by text describing the symbolic math process

Weaknesses/Questions:
I understand and like that it "works out of the box." It feels that authors focus more on correctness rather than performance. A few questions I have, are:
- What are the memory requirements of a "typical large enough" problem to be tackled?
- What are the advantages of DMP versus SMP only?
- Why are any strong scaling graphs not included?
- After reading the paper, I feel that MPI works correctly, but with no clue on why it is needed and what performance benefits it brings to the table

Minor:
- Have any of the additional targets been implemented in the meantime? Could the paper be updated there?
