### 2024-03-23 498e1f9
First updates based on reviewer comments
