+++ paper/ref.bib
@@ -328,16 +328,6 @@ series = {UAI'08}
   booktitle={Advances in Neural Information Processing Systems}
 }
 
-@ARTICLE{Syed2021NonReversible,
-  author = {Saifuddin Syed and Alexandre Bouchard-C\^{o}t\'{e} and George Deligiannidis and Arnaud Doucet},
-  title = {Non-reversible parallel tempering: a scalable highly parallel {MCMC} scheme},
-  journal = {Journal of Royal Statistical Society, Series B},
-  volume = {84},
-  pages = {321--350},
-  year = {2021}, 
-  doi = {10.1111/rssb.12464}
-}
-
 @article{lunn2000winbugs,
   title={Win{BUGS} --- a {B}ayesian modelling framework: concepts, structure, and extensibility},
   author={Lunn, David J and Thomas, Andrew and Best, Nicky and Spiegelhalter, David},
