+++ paper/examples.tex
@@ -3,6 +3,8 @@ In this section we present a set of minimal examples
 that demonstrate how to use Pigeons.jl for sampling. 
 For further reading, we also direct readers to our growing list of examples, which can be found at 
 \url{https://github.com/Julia-Tempering/Pigeons.jl/tree/main/examples}.
+\NS{Add InferenceHub description and link.}
+
 We begin by installing the latest official release of Pigeons.jl: 
 \begin{lstlisting}[language = Julia]
 using Pkg; Pkg.add("Pigeons")
