+++ paper/examples.tex
@@ -5,6 +5,7 @@ For further reading, we also direct readers to our growing list of examples, whi
 \url{https://github.com/Julia-Tempering/Pigeons.jl/tree/main/examples}.
 \NS{Add InferenceHub description and link.}
 
+\medskip
 We begin by installing the latest official release of Pigeons.jl: 
 \begin{lstlisting}[language = Julia]
 using Pkg; Pkg.add("Pigeons")
