### 2024-05-20 77f4601
Add weak scaling plots
