+++ paper/ref.bib
@@ -321,13 +321,6 @@ series = {UAI'08}
   year={2021}
 }
 
-@inproceedings{surjanovicparallel,
-  title={Parallel Tempering With a Variational Reference},
-  author={Surjanovic, Nikola and Syed, Saifuddin and Bouchard-C\^{o}t\'{e}, Alexandre and Campbell, Trevor},
-  year={2022},
-  booktitle={Advances in Neural Information Processing Systems}
-}
-
 @article{lunn2000winbugs,
   title={Win{BUGS} --- a {B}ayesian modelling framework: concepts, structure, and extensibility},
   author={Lunn, David J and Thomas, Andrew and Best, Nicky and Spiegelhalter, David},
@@ -712,4 +705,22 @@ archivePrefix = {arXiv},
   pages={1--11},
   year={2022},
   doi = {10.1145/3551624.3555300}
+}
+
+@inproceedings{syed2021paths,
+  title={Parallel tempering on optimized paths},
+  author={Syed, Saifuddin and Romaniello, Vittorio and Campbell, Trevor and Bouchard-C{\^o}t{\'e}, Alexandre},
+  booktitle={International Conference on Machine Learning},
+  pages={10033--10042},
+  year={2021},
+  organization={PMLR}
+}
+
+@inproceedings{biron2024automala,
+  title={auto{MALA}: {L}ocally adaptive {M}etropolis-adjusted {L}angevin algorithm},
+  author={Biron-Lattes, Miguel and Surjanovic, Nikola and Syed, Saifuddin and Campbell, Trevor and Bouchard-C{\^o}t{\'e}, Alexandre},
+  booktitle={International Conference on Artificial Intelligence and Statistics},
+  pages={4600--4608},
+  year={2024},
+  organization={PMLR}
 }
