### 2024-05-21 6e60663
Add new cite and fix typo
