+++ paper/ref.bib
@@ -724,3 +724,10 @@ archivePrefix = {arXiv},
   year={2024},
   organization={PMLR}
 }
+
+@article{surjanovic2024ergodicity,
+  title={Uniform ergodicity of parallel tempering with efficient local exploration},
+  author={Surjanovic, Nikola and Syed, Saifuddin and Bouchard-C\^{o}t\'{e}, Alexandre and Campbell, Trevor},
+  journal={arXiv:2405.11384},
+  year={2024}
+}
