+++ paper/PT.tex
@@ -352,10 +352,10 @@ We see that we obtain an approximately linear speedup as we increase the number 
         Weak scaling results for PT in both the distributed (MPI) and parallel settings 
         with approximate 95\% confidence intervals for the speedup.  
         The blue dotted line represents a hypothetical speedup equal to the 
-        number of processes/threads requested. 
+        number of processes/threads. 
         The red dashed line indicates a speedup of one for reference.
         \textbf{Left:} Distributed setup. 
-        \textbf{Right:} Parallel (shared) setup. 
+        \textbf{Right:} Parallel (shared memory) setup. 
     }
     \label{fig:PT_scaling}
 \end{figure*}
