+++ paper/examples.tex
@@ -97,8 +97,8 @@ model = coinflip(100000, 50000)
 
 From here, it is straightforward to sample from the density given by 
 \cref{eq:coinflip_density} up to a normalizing constant.
-We use non-reversible parallel tempering \NS{add newest ergodicity cite?} 
-\cite{syed2021nrpt,syed2021paths,surjanovic2022vpt} (PT), Pigeons.jl's 
+We use non-reversible parallel tempering 
+\cite{syed2021nrpt,syed2021paths,surjanovic2022vpt,surjanovic2024ergodicity} (PT), Pigeons.jl's 
 state-of-the-art sampling algorithm, to sample from the target distribution.
 PT comes with several tuning parameters and \cite{syed2021nrpt} describe how 
 to select these parameters effectively, which Pigeons.jl implements under the hood.
