### 2024-05-21 7975131
New tweaks based on Alex's comments
