+++ paper/PT.tex
@@ -308,8 +308,7 @@ with the updated permutation.
 
 \subsection{Weak scaling of PT}
 
-We study the weak scaling properties of Pigeons' implementation of PT
-as given by Gustafson's law.
+We study the weak scaling properties of Pigeons' implementation of PT.
 In this setup, we consider a varying problem size proportional to the number of processes used. 
 The number of chains is our definition of the problem size; 
 this is in line with previous heuristics that establish that the problem 
