### 2024-05-22 be25465
Update based on Trevor's comments
