+++ paper/examples.tex
@@ -304,9 +304,8 @@ The estimate can be accessed using
 stepping_stone(pt)
 \end{lstlisting}
 In the case of the normalization constant given by \cref{eq:coinflip_normalization} 
-for $n=100,000$ and $y=50,000$, we can exactly obtain its value as $\log(Z) \approx -11.8794$ 
-using a computer algebra system. Note that this is very close to the output provided in 
-\cref{fig:standard_out}.
+for $n=100,000$ and $y=50,000$, we can exactly obtain its value as $\log(Z) \approx -11.8794$.
+Note that this is very close to the output provided in \cref{fig:standard_out}.
 
 
 \subsubsection{Online statistics}
