### 2024-05-27 abc8ced
Starting to go through paper...
