+++ paper/paper.tex
@@ -35,7 +35,7 @@
 We introduce a software package, Pigeons.jl, that provides a way to leverage distributed computation to 
 obtain samples from complicated probability distributions, such as 
 multimodal posteriors arising in Bayesian inference and high-dimensional 
-distributions in statistical mechanics. Pigeons.jl provides simple APIs to 
+distributions in statistical mechanics. Pigeons.jl provides a simple user interface to 
 perform such computations single-threaded, multi-threaded,
 and/or distributed over thousands of MPI-communicating machines.
 In addition, Pigeons.jl guarantees a property that we call strong parallelism invariance: the 
