+++ paper/PT.tex
@@ -10,8 +10,6 @@ In this section we assume a basic understanding of
 Markov chain Monte Carlo (MCMC) methods.
 For readers unfamiliar with MCMC, it is important to know that 
 it is a method to obtain approximate samples from a distribution. 
-More specifically, it is a class of algorithms for simulating a Markov chain of 
-states that look like draws from the target distribution $\pi$.
 However, with traditional 
 MCMC methods, the samples may be heavily correlated instead of independent 
 and may fail to sufficiently explore the 
@@ -108,9 +106,9 @@ the top-right mode of the distribution presented in Figure~\ref{fig:bimodal}.
 
 
 \subsection{Local exploration and communication}
-A struct that is useful for the implementation of PT is the \texttt{Replica}, 
+A central data structure in Pigeons is the \texttt{Replica}, a Julia struct 
 which we will often refer to in the remainder of this section. 
-A single \texttt{Replica} struct stores a \texttt{state} variable and a \texttt{chain} 
+A single \texttt{Replica} stores a \texttt{state} variable and a \texttt{chain} 
 integer, among other entries. 
 At the beginning of PT, $N$ \texttt{Replicas} are created, 
 one for each distribution on the annealing path, and 
@@ -145,9 +143,9 @@ incurring a potentially very high communication cost for large values of $d$.
 
 
 \subsection{Distributed implementation}
-A distributed implementation of PT is presented in \cref{alg:distributed_PT}
+A distributed implementation of PT is shown in \cref{alg:distributed_PT}
 and we describe the details of the implementation below.
-We present the algorithm from the perspective that the number of machines available 
+For simplicity, we present the algorithm from the perspective that the number of machines available 
 is equal to the number of \texttt{Replica}s and distributions, $N$. 
 However, Pigeons.jl also allows for the more general case where the number of 
 machines is not necessarily equal to $N$.
@@ -156,9 +154,10 @@ machines is not necessarily equal to $N$.
 \label{sec:permuted_dist_array}
 Recall that our theoretical $O(1)$ message size for communication between two machines 
 is achieved by exchanging \texttt{chain} indices between \texttt{Replica}s 
-instead of their \texttt{state}s.
-One difficulty can be encountered in a distributed implementation, which we 
-illustrate with the following simple example. 
+instead of their \texttt{state}s. 
+However to achieve this low cost a specialized distributed data structure is 
+needed to maintain a mapping from chain indices to machine identifiers. 
+To motivate it, we start with an illustrative example. 
 Suppose we have $N=4$ distributions, chains, replicas, and machines, and that 
 machine 1 is exploring chain 2 while machine 4 is exploring chain 3.
 At one point, the \texttt{Replica} at \texttt{chain} 2 (machine 1) might need to exchange
@@ -187,9 +186,8 @@ We note that with a \texttt{PermutedDistributedArray} we make special assumption
 on how we access and write to the array elements. 
 Several MPI processes cooperate, with each machine storing 
 data for a slice of this distributed array, and at each time 
-step an index of the array is manipulated by exactly one machine.
-% With these assumptions we achieve read/write costs that are 
-% near-constant in the number of machines participating. 
+step an index of the array is manipulated by exactly one machine. 
+Thanks to this, we do not need to resort to global MPI synchronization steps. 
 
 \begin{figure*}[t] 
   \centering 
@@ -217,7 +215,7 @@ step an index of the array is manipulated by exactly one machine.
[…]
