+++ paper/PT.tex
@@ -145,8 +145,9 @@ incurring a potentially very high communication cost for large values of $d$.
 \subsection{Distributed implementation}
 A distributed implementation of PT is shown in \cref{alg:distributed_PT}
 and we describe the details of the implementation below.
-For simplicity, we present the algorithm from the perspective that the number of machines available 
-is equal to the number of \texttt{Replica}s and distributions, $N$. 
+For pedagogy, we first present the algorithm in the special case where the number of machines available 
+is equal to the number of \texttt{Replica}s and distributions, $N$. We outline 
+how Pigeons removes this restriction in \cref{sec:relax}.
 However, Pigeons.jl also allows for the more general case where the number of 
 machines is not necessarily equal to $N$.
 
@@ -252,21 +253,6 @@ convention, namely the shared sufficient statistic's uniform from the lowest
 interacting chain index) to make a common swap decision (where the 
 acceptance ratio is the sum of the two exchanged log ratios, (2)). 
 
-%In our pseudocode in \cref{alg:distributed_PT} we define the MPI-style 
-%functions \texttt{send()}, \texttt{receive!()}, 
-%and \texttt{waitall()}. 
-%The function \texttt{send()} has three arguments in our pseudocode: the first argument 
-%is the object to be sent, the second is the machine number to which the object should 
-%be sent, and finally the third argument is a unique tag for this specific send/receive 
-%request at this time step and on this machine, generated by \texttt{tag()}.
-%The \texttt{tag()} function can be any function that allows one to uniquely identify 
-%a tag for a given send/receive request on a given machine at any given time step $t$.
-%The function \texttt{receive!()} has the same last two arguments as \texttt{send()} 
-%except that the first argument 
-%specifies the object to which data should be written directly. Finally, 
-%\texttt{waitall(requests)} waits for all initialized MPI \texttt{requests} for a given pair 
-%of machines to complete.
-
 \medskip 
 In \cref{alg:distributed_PT} we also introduce the functions 
 \texttt{permuted\_get()} and \texttt{permuted\_set!()}, which operate 
@@ -286,20 +272,22 @@ with the updated permutation.
     \State $\texttt{dist\_array}[j] \gets j$ \Comment{\texttt{chain} $j$ is on machine $j$}
 		
     \For{$t$ {\bf in} 1, 2, \dots, $T$}
+    
+     \State $x_t \gets \texttt{local\_exploration}(\pi_\texttt{chain}, x_{t-1})$
+    \State $c_t \gets \texttt{chain}$ 
+    
 		  \If{$t$ is even} \Comment{choose between even or odd swap}
-		    \State $P \gets \{i: 1 \le i < N, i \text{ is even} \}$
+		    \State $\texttt{direction} \gets +1$
 		  \Else
-		    \State $P \gets \{i: 1 \le i < N, i \text{ is odd} \}$
+		    \State $\texttt{direction} \gets -1$
 		  \EndIf
+		  \State $\texttt{partner} \gets \texttt{chain} + \texttt{direction}$
 
-      \State $x_t \gets \texttt{local\_exploration}(\pi_\texttt{chain}, x_{t-1})$
-      \State $c_t \gets \texttt{chain}$ 
-      \If{$\texttt{chain} \in P$}
-        \State $\texttt{partner} \gets \texttt{chain}+1$
-      \ElsIf{$\texttt{chain}-1 \in P$}
-        \State $\texttt{partner} \gets \texttt{chain}-1$
-      \Else 
-      	\State $\texttt{partner} \gets \texttt{chain}$
+     
+      \If{$\texttt{partner} = 0$} 
[…]
