+++ paper/PT.tex
@@ -151,7 +151,7 @@ how Pigeons removes this restriction in \cref{sec:relax}.
 However, Pigeons.jl also allows for the more general case where the number of 
 machines is not necessarily equal to $N$.
 
-\subsubsection{The \texttt{PermutedDistributedArray}}
+\subsubsection{Distributed mapping from chains to machines}
 \label{sec:permuted_dist_array}
 Recall that our theoretical $O(1)$ message size for communication between two machines 
 is achieved by exchanging \texttt{chain} indices between \texttt{Replica}s 
