+++ paper/introduction.tex
@@ -42,7 +42,7 @@ Specifically, Pigeons.jl is designed to be suitable and yield reproducible outpu
 
 \subsection{Latest documentation}
 
-More in-depth documentation and additional examples can be found at: \\ \url{https://pigeons.run/}. 
+More in-depth documentation and additional examples can be found at: {\renewcommand{\UrlFont}{\ttfamily\normalsize}\url{https://pigeons.run/}}. 
 
 \subsection{Problem formulation}
 We describe the class of problems that can be approached using Pigeons.jl.
