+++ paper/paper.tex
@@ -31,7 +31,6 @@
 \maketitle
 
 \begin{abstract}
-
 We introduce a software package, Pigeons.jl, that provides a way to leverage distributed computation to 
 obtain samples from complicated probability distributions, such as 
 multimodal posteriors arising in Bayesian inference and high-dimensional 
@@ -43,7 +42,6 @@ output for a given seed is identical irrespective of the number of threads and p
 which is crucial for scientific reproducibility and software validation. 
 We describe the key features of Pigeons.jl and the approach taken to implement 
 a distributed and randomized algorithm that satisfies strong parallelism invariance.
-
 \end{abstract}
 
 \input{introduction.tex}
