+++ paper/introduction.tex
@@ -18,7 +18,7 @@ simulations that can last days or even weeks to finish.
 \medskip
 Pigeons.jl\footnote{The source code for Pigeons.jl can be found at 
 \url{https://github.com/Julia-Tempering/Pigeons.jl}. 
-Pigeons.jl v0.2.0, used in the example scripts below, is currently compatible with Julia 1.8+.}
+Pigeons.jl v0.2.0, used in the example scripts below, is compatible with Julia 1.8+.}
 enables users to sample efficiently from challenging probability distributions 
 and solve integration problems by 
 implementing state-of-the-art sampling algorithms \cite{syed2021nrpt,surjanovic2022vpt} 
@@ -46,8 +46,8 @@ More in-depth documentation and additional examples can be found at: {\renewcomm
 
 \subsection{Problem formulation}
 We describe the class of problems that can be approached using Pigeons.jl.
-Let $\pi(x)$ denote a probability density
-\footnote{This density may also be a probability mass function 
+Let $\pi(x)$ denote a probability density\footnote{This density may 
+also be a probability mass function 
 (discrete variables). We also allow for a combination of discrete/continuous variables.} 
 called the \emph{target}. 
 In many problems, e.g. in Bayesian statistics, the density $\pi$ is typically 
@@ -78,7 +78,7 @@ following scenarios:
     (e.g., high-dimensional, multi-modal, etc.).
     
     \item When the user needs not only $\int f(x) \pi(x) \, \dee x$ but also
-    the normalization constant $Z$. 
+    the normalizing constant $Z$. 
     Many existing tools focus on the former and struggle or fail to do the latter. 
     
     \item When the target distribution $\pi$ is defined over a non-standard space, 
