### 2024-05-29 #139 @nikola-sur
We thank the reviewers for their insightful comments! Our responses to each reviewer are presented below.

**Reviewer 1 (Jasan Barmparesos @nsailor)**
It would be very interesting to see the scaling characteristics of this implementation, especially with respect to the other algorithm parameters, for instance, the number of chains.
- Thank you for this comment. We have added weak scaling plots in the paper in the new section on weak scaling of PT (Section 3.4). In this context, the problem size is the number of PT chains. As the problem size increases, we also increase the number of processors for the weak scaling plots. For strong scaling plots, where the problem size (number of chains) is fixed, we propose a future addition to the Pigeons software package: “parallel parallel tempering”. Here, more than one copy of PT is run while keeping the total number of chains across all copies fixed. This parallel PT approach has been previously studied in the PT literature (Syed et al., 2022), but is not yet available in the current version of Pigeons.

Additionally, the paper mentions Pigeons being able to run on "thousands of MPI-communicating machines". It may be worth clarifying if this is something that has been tested or a possibility given the package's design.
- This has been tested (with up to 1000 MPI-communicating machines) on the computer clusters available at our institution. We have clarified this in the updated version of the paper.

I was not able to get a speedup with the toy MVN example using multiple threads (see [this issue](https://github.com/Julia-Tempering/Pigeons.jl/issues/191) in the project repository). In general, it would be very helpful to have some hints in the package's documentation for tuning the degree of parallelism for a given problem, especially if the speedup from parallelism is not linear.
- Thank you for raising this point. This issue has been closed as completed in the meantime. The reason for this is that the toy_mvn_target() has a special explorer performing iid sampling. We use this when we need very fast running tests (e.g. in CI) but in terms of performance profile this is atypical. Since the i.i.d. multivariate normal explorer is so quick the overhead of multithreading is not worth it. Typical explorers use algorithms such as slice sampling or gradient informed samplers, and for those algorithms we see improvements once the problem is large enough (and for small problems with fast explorers, absolute time spent will be small anyway, so relative speedup is less important). We have included new plots in the paper (added here for reference) comparing the speedup obtained with Pigeons and multi-threading in realistic scenarios.
![pigeons_multithread](https://github.com/JuliaCon/proceedings-review/assets/16364942/dfa81c9c-69d2-4528-8c0f-f47122b99f80)

It would be great to have additional examples, as described in [this issue](https://github.com/Julia-Tempering/Pigeons.jl/issues/179)
- We agree that this is a great addition. In fact, we have added a link in the paper to several real-life example use cases of Pigeons with these targets. In the meantime Pigeons has been applied to problems in physics, economics, astrostatistics, psychology, and more. Some examples are made publicly available on InferHub (https://julia-tempering.github.io/InferHub/), our new website dedicated to showcasing examples of applications of Pigeons. We plan to continue to grow this repository of examples over time.

**Reviewer 2 (George Bisbas @georgebisbas)**

What are the memory requirements of a "typical large enough" problem to be tackled?
- We emphasize that the memory requirements of a “difficult” sampling task are problem dependent. However, Pigeons has facilities to support cases requiring large memory. For instance, we allow for the computation of online statistics (https://pigeons.run/stable/output-online/), as well as off-memory sample storage (https://pigeons.run/stable/output-off-memory/). These have been commented on in the updated version of the manuscript.
- As a second remark, we note that Pigeons is intended for use with “complex” sampling problems. This does not necessarily mean that the memory requirements are large. For instance, a 100-dimensional real-valued distribution may have a very complicated structure, but only has a memory requirement of storing a floating point vector with 100 entries (800 bytes). Nevertheless, due to the complex nature of the problem, it may be necessary to use parallel tempering and distribute computation on up to several machines. In practice, the memory requirement is usually O(d), where d is the dimension of the target distribution. In the Bayesian setting, we may need to store n data points in order to evaluate the likelihood, in which case the memory requirement is O(d+n). Therefore, the memory requirements of a “difficult” sampling task are very much problem dependent.

What are the advantages of DMP versus SMP only?
- Thank you for this question. We note that Pigeons supports both shared memory parallelism (SMP) and distributed memory parallelism (DMP). In certain cases, it might be useful to run parallel tempering with very many chains (e.g., 100) or to run several instances of parallel tempering. In these cases, it is not reasonable to expect all users to have access to a shared memory resource capable of executing 100 processes in parallel. Our implementation in Pigeons ensures that the communication cost in the DMP setup is minimal. In fact, in most cases, the communication cost is negligible compared to other computations performed on each process. We have clarified these points in the paper. Additionally, DMP can be useful when the user's code is not thread-safe. In such cases, concurrency can still be achieved via DMP.

Why are any strong scaling graphs not included?
- Thank you for this comment. We have added weak scaling plots in the paper in the new section on weak scaling of PT (Section 3.4). In this context, the problem size is the number of PT chains. As the problem size increases, we also increase the number of processors for the weak scaling plots. For strong scaling plots, where the problem size (number of chains) is fixed, we propose a future addition to the Pigeons software package: “parallel parallel tempering”. Here, more than one copy of PT is run while keeping the total number of chains across all copies fixed. This parallel PT approach has been previously studied in the PT literature (Syed et al., 2022), but is not yet available in the current version of Pigeons. We are planning to implement this feature in the near future (https://github.com/Julia-Tempering/Pigeons.jl/issues/188).

After reading the paper, I feel that MPI works correctly, but with no clue on why it is needed and what performance benefits it brings to the table.
- Thank you for raising this point. We have added to the paper comments in a new section (Section 3.5) that emphasize that Pigeons allows for both DMP and SMP. In cases where the sampling problem is particularly complex, it can be advantageous to request many processes, in which case a distributed memory setup (e.g., on a cluster) is the most logical when it cannot be expected from the user to have access to a machine with, e.g., 100-1000 processors. MPI is the go-to interface for DMP settings that require communication between workers. Another notable example of the benefit of MPI is for the multi-GPU use case. For instance, in some situations the likelihood evaluation of a single chain requires its own GPU (e.g., for high-resolution black hole imaging, which relies on running a Fast Fourier Transform on a GPU).

Have any of the additional targets been implemented in the meantime? Could the paper be updated there?
- Yes, there are new targets, which are made publicly available on InferHub (https://julia-tempering.github.io/InferHub/), our new website dedicated to showcasing examples of applications of Pigeons. We plan to continue to grow this repository of examples over time. We have added a link in the paper to InferHub. In the meantime, Pigeons has also been applied to problems in physics, economics, astrostatistics, psychology, and more.

**References**
Syed, S., Bouchard-Côté, A., Deligiannidis, G., & Doucet, A. (2022). Non-reversible parallel tempering: a scalable highly parallel MCMC scheme. *Journal of the Royal Statistical Society Series B: Statistical Methodology*, 84(2), 321-350.
