### 2024-10-02 #139 @georgebisbas
Hi @pitsianis , @nikola-sur ,

I would expect that the editorial bot will create some new pdf. Is that right?
Is there any other way to check the updated manuscript?
