### 2024-10-02 #139 @nikola-sur
@georgebisbas The new PDF can be found above.
