### 2024-10-03 #139 @georgebisbas
Hi @nikola-sur , thanks for updating the manuscript.

I have updated my checklist above accordingly. I am more than happy with the response and the edits.

I think my only remaining checkboxes are:
> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

You may need to have a look here: https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors

>  References: Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?

I did not see DOIs in the software as far as I recall.

I think we are nearly there.
Best,
George
