### 2024-10-17 #139 @miguelbiron
Thank you both @nsailor @georgebisbas . Just to be clear on the contribution guidelines issue: this is referring to the repo, correct? Like having a proper CONTRIBUTING.md, for example. Not the paper itself, right?
