### 2024-10-18 #139 @georgebisbas
@miguelbiron , yes.
To the best of my understanding, this refers to the repo, and a CONTRIBUTING.md should be enough.
