+++ paper/ref.bib
@@ -1,710 +1,736 @@
-@article{bezanson2017julia,
-  title={Julia: A fresh approach to numerical computing},
-  author={Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B},
-  journal={SIAM review},
-  volume={59},
-  number={1},
-  pages={65--98},
-  year={2017},
-  publisher={SIAM}, 
-  DOI={10.1137/141000671}
+@article{10.1093/sysbio/syac033,
+    author = {Hartop, Emily and Srivathsan, Amrita and Ronquist, Fredrik and Meier, Rudolf},
+    doi = {10.1093/sysbio/syac033},
+    eprint = {https://academic.oup.com/sysbio/article-pdf/71/6/1404/46465829/syac033.pdf},
+    issn = {1063-5157},
+    journal = {Systematic Biology},
+    month = {05},
+    number = {6},
+    pages = {1404-1422},
+    title = {{Towards Large-Scale Integrative Taxonomy (LIT): Resolving the Data Conundrum for Dark Taxa}},
+    url = {https://doi.org/10.1093/sysbio/syac033},
+    volume = {71},
+    year = {2022}
 }
 
-@article{Moore1990,
-	author = {Moore, James T. and Bard, Jonathan F.},
-	issn = {0030-364X},
-	journal = {Operations Research},
-	keywords = {games: noncooperative,integer: branch-and-bound algorithms,programming},
-	mendeley-groups = {bilevel - hierarchical},
-	month = {oct},
-	number = {5},
-	pages = {911--921},
-	publisher = { INFORMS },
-	title = {{The Mixed Integer Linear Bilevel Programming Problem}},
-	url = {http://pubsonline.informs.org/doi/abs/10.1287/opre.38.5.911},
-	volume = {38},
-	year = {1990}
+@article{akiyama2021seven,
+    author = {Akiyama, Kazunori and Algaba, Juan Carlos and Alberdi, Antxon and Alef, Walter and Anantua, Richard and Asada, Keiichi and Azulay, Rebecca and Baczko, Anne-Kathrin and Ball, David and Balokovi{\'c}, Mislav and others},
+    journal = {The Astrophysical Journal Letters},
+    number = {1},
+    pages = {L12},
+    publisher = {IOP Publishing},
+    title = {{First M87 Event Horizon Telescope results. VII. Polarization of the ring}},
+    volume = {910},
+    year = {2021},
+	doi = {10.3847/2041-8213/abe71d}
 }
 
-@ARTICLE{syed2021nrpt,
-  author = {Saifuddin Syed and Alexandre Bouchard-C\^{o}t\'{e} and George Deligiannidis and Arnaud Doucet},
-  title = {Non-reversible parallel tempering: a scalable highly parallel {MCMC} scheme},
-  journal = {Journal of Royal Statistical Society, Series B},
-  volume = {84},
-  pages = {321--350},
-  year = {2021}
+@article{akiyama2022first,
+    author = {Akiyama, Kazunori and Alberdi, Antxon and Alef, Walter and Algaba, Juan Carlos and Anantua, Richard and Asada, Keiichi and Azulay, Rebecca and Bach, Uwe and Baczko, Anne-Kathrin and Ball, David and others},
+    journal = {The Astrophysical Journal Letters},
+    number = {2},
+    pages = {L15},
+    publisher = {IOP Publishing},
+    title = {First {S}agittarius {A}* {E}vent {H}orizon {T}elescope results. {IV}. {V}ariability, morphology, and black hole mass},
+    volume = {930},
+    year = {2022},
+	doi = {10.3847/2041-8213/ac6736}
 }
 
[…]
