### 2024-11-07 #139 @pitsianis
Indeed very nice work and paper. Unrelated, it would be nice to have `slurm` commands to submit to a job queue. There exist some very nice examples where an otherwise normal `.jl` script can also act as a job submission script. That coupled with `DrWatson.jl` can build a powerful environment to run and rerun heavy computations.
