### 2024-11-07 #139 @pitsianis
@nikola-sur please have another look at the DOIs of your references, see the report above.
Also, see the post-review checklist above, and post here your answers for every bullet point, so that I can check the remaining of the checkboxes as well.
