### 2024-11-26 cbd0857
Update ref.bib
