+++ paper/ref.bib
@@ -187,9 +187,10 @@
     number = {4},
     pages = {327--335},
     publisher = {Taylor \& Francis},
-    title = {Understanding the {M}etropolis-{H}astings algorithm},
+    title = {Understanding the {M}etropolis--{H}astings algorithm},
     volume = {49},
-    year = {1995}
+    year = {1995}, 
+    doi = {10.1080/00031305.1995.10476177}
 }
 
 @inproceedings{cusumano2019gen,
@@ -243,7 +244,8 @@
     publisher = {Taylor \& Francis},
     title = {Gibbs sampling},
     volume = {95},
-    year = {2000}
+    year = {2000}, 
+    doi = {10.1080/01621459.2000.10474335}
 }
 
 @article{geweke2004getting,
@@ -281,7 +283,8 @@
     publisher = {AUAI Press},
     series = {UAI'08},
     title = {Church: a Language for Generative Models},
-    year = {2008}
+    year = {2008}, 
+    doi = {10.5555/3023476.3023503}
 }
 
 @article{gota2021overview,
@@ -302,7 +305,8 @@
     publisher = {Europe PMC Funders},
     title = {MultiBUGS: a parallel implementation of the BUGS modelling framework for faster {B}ayesian inference},
     volume = {95},
-    year = {2020}
+    year = {2020}, 
+    doi = {10.18637/jss.v095.i07}
 }
 
 @article{grun2022many,
@@ -313,7 +317,8 @@
     publisher = {Springer},
     title = {How many data clusters are in the Galaxy data set? Bayesian cluster analysis in action},
     volume = {16},
-    year = {2022}
+    year = {2022}, 
+    doi = {10.1007/s11634-021-00461-8}
 }
 
 @article{hoffman2014no,
@@ -323,7 +328,8 @@
     pages = {1593--1623},
     title = {The No-U-Turn sampler: adaptively setting path lengths in Hamiltonian Monte Carlo.},
     volume = {15},
-    year = {2014}
+    year = {2014}, 
+    doi = {10.5555/2627435.2638586}
 }
 
 @inproceedings{hornik2003jags,
@@ -509,7 +515,8 @@
     publisher = {Springer},
     title = {The {B}ayesian Choice: From Decision-Theoretic Foundations to Computational Implementation},
     volume = {2},
-    year = {2007}
+    year = {2007}, 
+    doi = {10.1007/0-387-71599-1}
 }
 
 @article{roeder1990density,
