### 2024-11-26 #139 @nikola-sur
Hello @pitsianis, we have gone through your checklist and have tried our best to address as many of the DOIs above. Here are answers to your questions:

  - Checked.
- Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JuliaCon Proceedings paper.
  - Pigeons v0.4.8.
  - 10.5281/zenodo.14226593
- Make sure that the title and author list (including ORCIDs) in the archive match those in the JuliaCon Proceedings paper.
  - Checked.
  - Checked.
