### 2025-02-28 #139 @GiggleLiu
## Summary
This work implements several methods to find the (close to) optimal contraction paths for tensor networks. The benchmark shows the advantage in the run time. Tensor network is a modern linear algebra, and contraction paths is a fundamental concept in this field. Finding better contraction paths can greatly improve the contraction of tensor networks.

## Comments
1. I hope the authors can clarify more on the what are the new contribution to the community. For example, I would expect the authors to demonstrate an algorithm that not yet implemented in the past.

2. About the writing, I doubt the manuscript is not ready for review yet:
    1. The title "extttEinExprs.jl: Contraction Paths as Symbolic Expressions" has an extra "exttt" before it. Maybe this is a typo?
    2. In the abstract, the authors include a incomplete sentense: "We benchmark this library against ... with random tensor networks."

3. As the main result, the authors show the benchmark of run-time of contraction order finder with existing packages: `OMEinsumContractionOders` and `TensorOperations`. I can see 2 issues.
    1. I think the resulting tree-width of a contraction order should be the key metric of contraction order optimizer, which is not shown here.
    2. In `OMEinsumContractionOders`,  the greedy algorthm is repeatedly run for 10 times by default. I hope the authors can open source the benchmark repo as well so that I can help them check if the parameters are setup properly.

I do not think this work should be accepted in the current form. Major modification is needed. The authors should implement a new algorithm that not yet available in this community, benchmark and show it can beat others with lower tree-width, and improve the writing.
