### 2025-03-04 5cfe997
Fix missing DOIs etc
