+++ paper/ref.bib
@@ -79,8 +79,7 @@
     publisher = {JMLR. org},
     title = {On {M}arkov chain {M}onte {C}arlo methods for tall data},
     volume = {18},
-    year = {2017},
-	doi = {10.5555/3122009.3153003}
+    year = {2017}
 }
 
 @article{bezanson2017julia,
@@ -268,7 +267,8 @@
     publisher = {Wiley Online Library},
     title = {A language and program for complex {B}ayesian modelling},
     volume = {43},
-    year = {1994}
+    year = {1994}, 
+    doi = {10.2307/2348941}
 }
 
 @inproceedings{goodman2008church,
@@ -283,8 +283,7 @@
     publisher = {AUAI Press},
     series = {UAI'08},
     title = {Church: a Language for Generative Models},
-    year = {2008}, 
-    doi = {10.5555/3023476.3023503}
+    year = {2008}
 }
 
 @article{gota2021overview,
@@ -328,8 +327,7 @@
     pages = {1593--1623},
     title = {The No-U-Turn sampler: adaptively setting path lengths in Hamiltonian Monte Carlo.},
     volume = {15},
-    year = {2014}, 
-    doi = {10.5555/2627435.2638586}
+    year = {2014}
 }
 
 @inproceedings{hornik2003jags,
@@ -348,7 +346,8 @@
     publisher = {Oxford University Press},
     title = {MRBAYES: {B}ayesian inference of phylogenetic trees},
     volume = {17},
-    year = {2001}
+    year = {2001}, 
+    doi = {10.1093/bioinformatics/17.8.754}
 }
 
 @article{jacob2011using,
@@ -383,7 +382,8 @@
     pages = {69--93},
     title = {{H}amiltonian {M}onte {C}arlo in inverse problems; Ill-conditioning and multimodality},
     volume = {13},
-    year = {2023}
+    year = {2023}, 
+    doi = {10.1615/int.j.uncertaintyquantification.2022038478}
 }
 
 @article{lartillot2013phylobayes,
@@ -507,7 +507,8 @@
     title = {The {ngEHT}'s Role in Measuring Supermassive Black Hole Spins},
     url = {https://www.mdpi.com/2075-4434/11/1/6},
     volume = {11},
-    year = {2023}
+    year = {2023}, 
+    doi = {10.3390/galaxies11010006}
 }
 
 @book{robert2007bayesian,
@@ -585,7 +586,8 @@
     pages = {2741},
     title = {Para{M}onte: A high-performance serial/parallel {M}onte {C}arlo simulation library for {C}, {C++}, {F}ortran},
     volume = {6},
-    year = {2021}
+    year = {2021}, 
+    doi = {10.21105/joss.02741}
 }
 
 @article{steele2014fast,
@@ -604,8 +606,7 @@
     author = {Surjanovic, Nikola and Syed, Saifuddin and Bouchard-C\^{o}t\'{e}, Alexandre and Campbell, Trevor},
     booktitle = {Advances in Neural Information Processing Systems},
     title = {Parallel Tempering With a Variational Reference},
-    year = {2022},
-	doi = {10.5555/3600270.3600311}
+    year = {2022}
 }
 
 @inproceedings{surjanovicparallel,
@@ -621,7 +622,8 @@
     organization = {IEEE},
     pages = {1--6},
     title = {Large scale {B}ayesian data analysis for nuclear fusion experiments},
-    year = {2007}
+    year = {2007}, 
+    doi = {10.1109/wisp.2007.4447579}
 }
 
 @article{swendsen1986replica,
